Remove IntoPII source code from 3rdparth folder
Remove IntoPII from 3rd Party library, don't use it anymore, has adapted
to OpenCV for a while, and quit satisfice with it.
IntoPII don't have many documents, all operations are hidden in streams
and call backs, which makes harder to trace and manipulate directly on
the raw data. The idea of using Qt and streaming are kind of smart
though.